requirement.txt
numpy
pandas
matplotlib
flask
torch
torchvision
gradio
accelerate
seaborn
scikit-learn
pillow

